refactor: typescript icons
