+++ src/components/relicPreview/GenerateStat.tsx
@@ -2,12 +2,12 @@ import { Flex } from 'antd'
 import { Renderer } from 'lib/renderer'
 import { Assets } from 'lib/assets'
 import { Utils } from 'lib/utils'
-import { Constants, SubStats } from 'lib/constants'
+import { SubStats } from 'lib/constants'
 import { iconSize } from 'lib/constantsUi'
 import RelicStatText from 'components/relicPreview/RelicStatText'
 import { Relic, StatRolls } from 'types/Relic'
 import { ReactElement } from 'react'
-import { RightIcon } from 'icons/RightIcon.jsx'
+import { RightIcon } from 'icons/RightIcon'
 import { useTranslation } from 'react-i18next'
 
 type Substat = {
